Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-remove-source-after-expected.txt


FAIL pointer updates (removing source element after pointer) assert_equals: error events on c expected 1 but got 0
